Bootstrap dev site entries to allow for chunking
